feat: Make dart json generators explicitly call toJson()
When passing a class to jsonEncode() in dart it will traverse the json
tree calling toJson() on all of the nested attributes. If, however, code
calls toJson() on a class and expects it to be completely serialized it
will fail because toJson() is not called on the nested attributes.

This commit sets the explicitToJson flag to true to fix this.